refactor: share provider credential extraction logic
