+++ src-tauri/src/services/provider/mod.rs
@@ -929,6 +929,108 @@ base_url = "http://localhost:8080"
     }
 }
 
+pub(super) fn non_empty_trimmed(value: Option<&str>) -> Option<String> {
+    value
+        .map(str::trim)
+        .filter(|v| !v.is_empty())
+        .map(ToOwned::to_owned)
+}
+
+pub(super) fn extract_toml_base_url(config_toml: &str) -> Option<String> {
+    let re = Regex::new(r#"base_url\s*=\s*["']([^"']+)["']"#).ok()?;
+    re.captures(config_toml)
+        .and_then(|caps| caps.get(1))
+        .map(|m| m.as_str().to_string())
+}
+
+pub(super) fn extract_provider_api_key(provider: &Provider, app_type: &AppType) -> Option<String> {
+    let settings = &provider.settings_config;
+
+    let direct = match app_type {
+        AppType::Codex => non_empty_trimmed(
+            settings
+                .get("auth")
+                .and_then(|v| v.get("OPENAI_API_KEY"))
+                .and_then(|v| v.as_str()),
+        ),
+        _ => None,
+    };
+
+    direct
+        .or_else(|| {
+            non_empty_trimmed(
+                settings
+                    .get("apiKey")
+                    .or_else(|| settings.get("api_key"))
+                    .and_then(|v| v.as_str())
+                    .or_else(|| {
+                        settings
+                            .get("options")
+                            .and_then(|v| v.get("apiKey").or_else(|| v.get("api_key")))
+                            .and_then(|v| v.as_str())
+                    }),
+            )
+        })
+        .or_else(|| {
+            let env = settings.get("env")?;
+            non_empty_trimmed(
+                env.get("ANTHROPIC_AUTH_TOKEN")
+                    .or_else(|| env.get("ANTHROPIC_API_KEY"))
+                    .or_else(|| env.get("OPENAI_API_KEY"))
+                    .or_else(|| env.get("OPENROUTER_API_KEY"))
+                    .or_else(|| env.get("GOOGLE_API_KEY"))
+                    .or_else(|| env.get("GEMINI_API_KEY"))
+                    .or_else(|| env.get("CODEX_API_KEY"))
+                    .and_then(|v| v.as_str()),
+            )
+        })
+}
+
+pub(super) fn extract_provider_base_url(
+    provider: &Provider,
+    app_type: &AppType,
+) -> Option<String> {
+    let settings = &provider.settings_config;
+
+    let direct = match app_type {
+        AppType::Codex => settings
+            .get("config")
+            .and_then(|v| v.as_str())
+            .and_then(extract_toml_base_url),
+        _ => None,
+    };
+
+    direct
+        .or_else(|| {
+            non_empty_trimmed(
+                settings
+                    .get("baseUrl")
+                    .or_else(|| settings.get("baseURL"))
+                    .or_else(|| settings.get("base_url"))
+                    .and_then(|v| v.as_str())
+                    .or_else(|| {
+                        settings
+                            .get("options")
+                            .and_then(|v| {
+                                v.get("baseURL")
+                                    .or_else(|| v.get("baseUrl"))
+                                    .or_else(|| v.get("base_url"))
+                            })
+                            .and_then(|v| v.as_str())
+                    }),
+            )
+        })
+        .or_else(|| {
+            let env = settings.get("env")?;
+            non_empty_trimmed(
+                env.get("ANTHROPIC_BASE_URL")
+                    .or_else(|| env.get("GOOGLE_GEMINI_BASE_URL"))
+                    .or_else(|| env.get("OPENAI_BASE_URL"))
+                    .and_then(|v| v.as_str()),
+            )
+        })
+}
+
 impl ProviderService {
     fn normalize_provider_if_claude(app_type: &AppType, provider: &mut Provider) {
         if matches!(app_type, AppType::Claude) {
@@ -2096,110 +2198,48 @@ impl ProviderService {
     ) -> Result<(String, String), AppError> {
         match app_type {
             AppType::Claude => {
-                let env = provider
-                    .settings_config
-                    .get("env")
-                    .and_then(|v| v.as_object())
-                    .ok_or_else(|| {
-                        AppError::localized(
-                            "provider.claude.env.missing",
-                            "配置格式错误: 缺少 env",
-                            "Invalid configuration: missing env section",
-                        )
-                    })?;
-
-                let api_key = env
-                    .get("ANTHROPIC_AUTH_TOKEN")
-                    .or_else(|| env.get("ANTHROPIC_API_KEY"))
-                    .and_then(|v| v.as_str())
+                let api_key = extract_provider_api_key(provider, app_type)
                     .ok_or_else(|| {
                         AppError::localized(
                             "provider.claude.api_key.missing",
                             "缺少 API Key",
                             "API key is missing",
                         )
-                    })?
-                    .to_string();
+                    })?;
 
-                let base_url = env
-                    .get("ANTHROPIC_BASE_URL")
-                    .and_then(|v| v.as_str())
+                let base_url = extract_provider_base_url(provider, app_type)
                     .ok_or_else(|| {
                         AppError::localized(
                             "provider.claude.base_url.missing",
                             "缺少 ANTHROPIC_BASE_URL 配置",
                             "Missing ANTHROPIC_BASE_URL configuration",
                         )
-                    })?
-                    .to_string();
+                    })?;
 
                 Ok((api_key, base_url))
             }
             AppType::Codex => {
-                let auth = provider
-                    .settings_config
-                    .get("auth")
-                    .and_then(|v| v.as_object())
-                    .ok_or_else(|| {
-                        AppError::localized(
-                            "provider.codex.auth.missing",
-                            "配置格式错误: 缺少 auth",
-                            "Invalid configuration: missing auth section",
-                        )
-                    })?;
-
-                let api_key = auth
-                    .get("OPENAI_API_KEY")
-                    .and_then(|v| v.as_str())
+                let api_key = extract_provider_api_key(provider, app_type)
                     .ok_or_else(|| {
                         AppError::localized(
                             "provider.codex.api_key.missing",
                             "缺少 API Key",
                             "API key is missing",
                         )
-                    })?
-                    .to_string();
-
-                let config_toml = provider
-                    .settings_config
-                    .get("config")
-                    .and_then(|v| v.as_str())
-                    .unwrap_or("");
-
-                let base_url = if config_toml.contains("base_url") {
-                    let re = Regex::new(r#"base_url\s*=\s*["']([^"']+)["']"#).map_err(|e| {
-                        AppError::localized(
-                            "provider.regex_init_failed",
-                            format!("正则初始化失败: {e}"),
-                            format!("Failed to initialize regex: {e}"),
-                        )
                     })?;
-                    re.captures(config_toml)
-                        .and_then(|caps| caps.get(1))
-                        .map(|m| m.as_str().to_string())
-                        .ok_or_else(|| {
-                            AppError::localized(
-                                "provider.codex.base_url.invalid",
-                                "config.toml 中 base_url 格式错误",
-                                "base_url in config.toml has invalid format",
-                            )
-                        })?
-                } else {
-                    return Err(AppError::localized(
+
+                let base_url = extract_provider_base_url(provider, app_type).ok_or_else(|| {
+                    AppError::localized(
                         "provider.codex.base_url.missing",
                         "config.toml 中缺少 base_url 配置",
                         "base_url is missing from config.toml",
-                    ));
-                };
+                    )
+                })?;
 
                 Ok((api_key, base_url))
             }
             AppType::Gemini => {
-                use crate::gemini_config::json_to_env;
-
-                let env_map = json_to_env(&provider.settings_config)?;
-
-                let api_key = env_map.get("GEMINI_API_KEY").cloned().ok_or_else(|| {
+                let api_key = extract_provider_api_key(provider, app_type).ok_or_else(|| {
                     AppError::localized(
                         "gemini.missing_api_key",
                         "缺少 GEMINI_API_KEY",
@@ -2207,68 +2247,36 @@ impl ProviderService {
                     )
                 })?;
 
-                let base_url = env_map
-                    .get("GOOGLE_GEMINI_BASE_URL")
-                    .cloned()
+                let base_url = extract_provider_base_url(provider, app_type)
                     .unwrap_or_else(|| "https://generativelanguage.googleapis.com".to_string());
 
                 Ok((api_key, base_url))
             }
             AppType::OpenCode => {
-                // OpenCode uses options.apiKey and options.baseURL
-                let options = provider
-                    .settings_config
-                    .get("options")
-                    .and_then(|v| v.as_object())
-                    .ok_or_else(|| {
-                        AppError::localized(
-                            "provider.opencode.options.missing",
-                            "配置格式错误: 缺少 options",
-                            "Invalid configuration: missing options section",
-                        )
-                    })?;
-
-                let api_key = options
-                    .get("apiKey")
-                    .and_then(|v| v.as_str())
+                let api_key = extract_provider_api_key(provider, app_type)
                     .ok_or_else(|| {
                         AppError::localized(
                             "provider.opencode.api_key.missing",
                             "缺少 API Key",
                             "API key is missing",
                         )
-                    })?
-                    .to_string();
+                    })?;
 
-                let base_url = options
-                    .get("baseURL")
-                    .and_then(|v| v.as_str())
-                    .unwrap_or("")
-                    .to_string();
+                let base_url = extract_provider_base_url(provider, app_type).unwrap_or_default();
 
                 Ok((api_key, base_url))
             }
             AppType::OpenClaw => {
-                // OpenClaw uses apiKey and baseUrl directly on the object
-                let api_key = provider
-                    .settings_config
-                    .get("apiKey")
-                    .and_then(|v| v.as_str())
+                let api_key = extract_provider_api_key(provider, app_type)
                     .ok_or_else(|| {
                         AppError::localized(
                             "provider.openclaw.api_key.missing",
                             "缺少 API Key",
                             "API key is missing",
                         )
-                    })?
-                    .to_string();
+                    })?;
 
-                let base_url = provider
-                    .settings_config
-                    .get("baseUrl")
-                    .and_then(|v| v.as_str())
-                    .unwrap_or("")
-                    .to_string();
+                let base_url = extract_provider_base_url(provider, app_type).unwrap_or_default();
 
                 Ok((api_key, base_url))
             }
